requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
scikit-learn==1.4.2
nltk==3.6.5
spacy>=3.7.2,<4.0.0
psycopg2-binary==2.9.1
sqlalchemy==1.4.31
tensorflow>=2.6.0
transformers==4.40.1
PyAudio==0.2.14
SpeechRecognition==3.10.4
sqlalchemy>=1.4.22
